fix issues interface
