+++ src/typeScript/app/app.ts
@@ -4,7 +4,7 @@ import AppView from "../views/appView";
 import AuthController from "../controllers/authController";
 import AuthService from "../services/auth.service";
 import AuthView from "../views/authView";
-import { IAuthParam } from "../constants/interface";
+import { IAuthParam } from "../interfaces";
 
 const authParam: IAuthParam = {
   AuthService: new AuthService(),
