+++ src/typeScript/interfaces/index.ts
@@ -1,0 +1,4 @@
+import { IAuth, IAuthParam } from "./auth";
+import { ITodo, ITodoParam } from "./todo";
+
+export type { IAuth, IAuthParam, ITodo, ITodoParam };
