+++ src/typeScript/services/todo.service.ts
@@ -4,7 +4,7 @@ import AppView from "../views/appView";
 import TodoSchema from "../models/todoModel";
 import TOAST from "../helper/toast";
 import MESSAGE from "../constants/message";
-import { ITodo } from "../constants/interface";
+import { ITodo } from "../interfaces";
 import { TodoData } from "../constants/types";
 
 class TodoService {
